move func to separate file
