+++ src/app/api/inngest/route.ts
@@ -1,6 +1,7 @@
 import { serve } from "inngest/next";
 import { inngest } from "../../../inngest/client";
-import { inngest_downloadAndExtractTextFromFile } from "../webhooks/on-download-complete/route";
+import { inngest_downloadAndExtractTextFromFile } from "../webhooks/on-download-complete/downloadAndExtractTextFromFile";
+
 
 export const { GET, POST, PUT } = serve({
   client: inngest,
